# api-doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Tripjoy API Service Documentation",
    "description": "This API exposes Tripjoy endpoints.",
    "contact": {
      "name": "Nguyen Dinh Duc",
      "url": "https://www.linkedin.com/in/dinhduc0794",
      "email": "dinhduc0794.work@gmail.com"
    },
    "license": {
      "name": "API License",
      "url": "https://springdoc.org"
    },
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Environment"
    },
    {
      "url": "https://api.tripjoy.com",
      "description": "Production Environment"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/groups/{groupId}": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Get group information by ID - OK",
        "operationId": "getGroupById",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseGroupResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Groups"
        ],
        "summary": "Update group information - OK",
        "operationId": "updateGroup",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseGroupResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Groups"
        ],
        "summary": "Delete group (soft delete with cascade) - Only LEADER",
        "operationId": "deleteGroup",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{groupId}/members/{memberId}": {
      "put": {
        "tags": [
          "Groups"
        ],
        "summary": "Update member role (Leader/CO_LEADER/Member) - OK",
        "operationId": "updateMemberRole",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseGroupMemberResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Groups"
        ],
        "summary": "Remove member from group (Automatically kicks from Chat) - OK",
        "operationId": "removeMember",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/groups": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Get my groups list - OK",
        "operationId": "getMyGroups",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListGroupResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Create a new group (Automatically creates General chat channel) - OK",
        "operationId": "createGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseGroupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{groupId}/transfer-leadership": {
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Transfer leadership to another member - OK",
        "operationId": "transferLeadership",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferLeadershipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{groupId}/members": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Get group members list - OK",
        "operationId": "getMembers",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListGroupMemberResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Add member to group (Automatically syncs to Chat) - OK",
        "operationId": "addMember",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{groupId}/location-suggestions": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Get list of suggested locations",
        "operationId": "getSuggestedLocations",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListSuggestLocationResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Suggest location for group",
        "operationId": "suggestLocation",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestLocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSuggestLocationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/groups/{groupId}/members/me": {
      "delete": {
        "tags": [
          "Groups"
        ],
        "summary": "Leave group (Self-initiated) - OK",
        "operationId": "leaveGroup",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "UserSimpleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          }
        }
      },
...
      "LocationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lng": {
            "type": "number",
            "format": "double"
          },
          "hotline": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "isOpen": {
            "type": "boolean"
          },
          "content": {
            "type": "string"
          }
        }
      },
...
      "GroupRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the travel group",
            "example": "Da Nang Avengers"
          },
          "avatar": {
            "type": "string",
            "description": "URL of the group's avatar image",
            "example": "https://res.cloudinary.com/tripjoy/image/upload/group_avatar.jpg"
          },
          "description": {
            "type": "string",
            "description": "Description of the travel group"
          },
          "chatbotCount": {
            "type": "integer",
            "format": "int32"
          },
          "isPro": {
            "type": "boolean"
          },
          "theme_color": {
            "type": "string",
            "description": "Hex color code for group theme",
            "example": "#FF5733"
          },
          "member_ids": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "ApiResponseGroupResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/GroupResponse"
          }
        }
      },
      "GroupMemberResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/UserSimpleResponse"
          },
          "role": {
            "type": "string",
            "enum": [
              "LEADER",
              "CO_LEADER",
              "MEMBER"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by": {
            "type": "string",
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_by": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "GroupResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "theme": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "members": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupMemberResponse"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by": {
            "type": "string",
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_by": {
            "type": "string",
            "format": "uuid"
          },
          "theme_color": {
            "type": "string"
          },
          "is_pro": {
            "type": "boolean"
          },
          "chatbot_count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "UpdateMemberRoleRequest": {
        "required": [
          "role"
        ],
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "description": "Assign new role to member",
            "example": "CO_LEADER",
            "enum": [
              "LEADER",
              "CO_LEADER",
              "MEMBER"
            ]
          }
        }
      },
      "ApiResponseGroupMemberResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/GroupMemberResponse"
          }
        }
      },
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "TransferLeadershipRequest": {
        "required": [
          "newLeaderId"
        ],
        "type": "object",
        "properties": {
          "newLeaderId": {
            "type": "string",
            "description": "ID of the member to become the new leader",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        }
      },
      "AddMemberRequest": {
        "required": [
          "member_id",
          "role"
        ],
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "description": "Assign new role to member",
            "example": "CO_LEADER",
            "enum": [
              "LEADER",
              "CO_LEADER",
              "MEMBER"
            ]
          },
          "member_id": {
            "type": "string",
            "description": "UUID of the users to be added to the group",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
          }
        }
      },
      "SuggestLocationRequest": {
        "required": [
          "location_id"
        ],
        "type": "object",
        "properties": {
          "location_id": {
            "type": "string",
            "description": "The UUID of the Location being suggested",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
          },
          "notes": {
            "type": "string",
            "description": "Optional notes about why this location is suggested",
            "example": "This place looks great for our first day!"
          }
        }
      },
      "ApiResponseSuggestLocationResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/SuggestLocationResponse"
          }
        }
      },
      "SuggestLocationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "location": {
            "$ref": "#/components/schemas/LocationResponse"
          },
          "notes": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by": {
            "type": "string",
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_by": {
            "type": "string",
            "format": "uuid"
          },
          "suggested_by": {
            "$ref": "#/components/schemas/UserSimpleResponse"
          },
          "group_id": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseListGroupResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupResponse"
            }
          }
        }
      },
      "ApiResponseListGroupMemberResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupMemberResponse"
            }
          }
        }
      },
      "ApiResponseListSuggestLocationResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SuggestLocationResponse"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}